// src/types/projects.ts
export type TProject = {
  _id: string;
  title: string;
  thumbnail: string;
  description: string;
  techStack: string[];
  duration: string;
  team: string;
  category: string;
  status: string;
  features:string[];
  challenges: string[];
  gallery:string[]
  frontendLivelink: string;
  backendLivelink: string;
  githubFrontend: string;
  githubBackend: string;
  githubFullStack: string;
  problemFaced: string;
};